Fixes issue #7 "RuntimeException: Could not open table" was occurring when you run the DbBenchmark
Active version objects are held in a weak hash map.  Once they are no longer in the map, then their associated file are candidates to be deleted.  The problem is that open Snapshots and Iterators were not holding to a reference of the Version object which they were associated with.  This resulted in files associated with the version being deleted while it was still being iterated.

Also added Version reference counting so that if the end user is properly closing the snapshots and iterators, then Version will be evicted from active versions map.  This should allow faster reclaiming of disk space even in situations where the JVM is not eagerly GCing.