When deleting obsolete files, ignore file name patterns we don't understand.
This change avoids a NullPointerException being thrown in cases where
a file is present in the LevelDB database directory whose name does
not match any of the patterns detected by Filename#parseFileName().
Such files may be placed in the directory by the embedding
application.

Note that the C++ LevelDB implementation already ignores such foreign
files, per how its leveldb::DBImpl::DeleteObsoleteFiles() function
calls on the leveldb::ParseFileName() function, which is documented to
return false if the given file name is not "a LevelDB file."