Fix SnapshotSeekingIterator and compaction issues with deleted keys
SnapshotSeekingIterator could return deleted entries (it should't) depending on the order they where disposed in source iterator.
File selection in compaction process was not collecting all overlapping files in level 0. This could cause the db to forget that one entry was deleted after a compaction (tombstone removed, but not the earlier entry).