Fix an issue where a large byte array may get pinned down in memory which could lead to out of memory.
Why:
FileMetadata stores smallest and largest key internally to identity
the start and end key range in a given file. The key is essentially
a slice of a byte array storing a key-value entry. This causes
the byte array to be pinned down in memory, which can be an issue
when the value is fairly large. Eventually this can lead to
out of memory.

What:
The change is to make a copy of the key (which should be typically
small) into a new byte array so the original byte array
(which can be large) can be freed.